fix: rename NavbarComponent.svelte to NavBar.svelte for consistency
