+++ src/lib/components/NavBar.svelte
